prueba cambio long

+++ src/main/java/com/conexion/api/repository/ProductosRepository.java
@@ -4,4 +4,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.conexion.api.model.Productos;
 
-public interface ProductosRepository extends JpaRepository<Productos, Integer> {}
+public interface ProductosRepository extends JpaRepository<Productos, Long> {}
